put tcp stuff in sharedNetworking
