+++ CSharpEindopdracht/Client/Client.cs
@@ -1,52 +1,22 @@
-﻿using System;
+﻿using SharedNetworking.Utils;
+using System;
 using System.Collections.Generic;
 using System.Net.Sockets;
 using System.Text;
 
 namespace Server
 {
-    internal class Client
+    internal class Client : SharedNetworking.Utils.SharedClient
     {
-        private TcpClient client;
-        private NetworkStream stream;
-        private byte[] buffer = new byte[1024];
-        private byte[] totalBuffer = new byte[1024];
-        private int totalBufferReceived = 0;
+
         private string username;
         private int clientID;
 
-        public Client(TcpClient tcpClient, int clientID)
+        public Client(TcpClient tcpClient, DataHandler dataHandler, int clientID) : base(tcpClient, dataHandler)
         {
             this.clientID = clientID;
         }
 
-        private void OnRead(IAsyncResult ar)
-        {
-            int receivedBytes = this.stream.EndRead(ar);
-
-            if (totalBufferReceived + receivedBytes > buffer.Length)
-            {
-                throw new OutOfMemoryException("buffer too small");
-            }
-            Array.Copy(buffer, 0, totalBuffer, totalBufferReceived, receivedBytes);
-            totalBufferReceived += receivedBytes;
-
-            ushort expectedMessageLength = BitConverter.ToUInt16(totalBuffer, 0);
-            while (totalBufferReceived >= expectedMessageLength)
-            {
-                //volledig packet binnen
-                byte[] messageBytes = new byte[expectedMessageLength];
-                Array.Copy(totalBuffer, 0, messageBytes, 0, expectedMessageLength);
-
-
-                //byte[] payloadbytes = new byte[BitConverter.ToInt32(messageBytes, 0) - 5];
-
-                //Array.Copy(messageBytes, 5, payloadbytes, 0, payloadbytes.Length);
-
-            }
-
-            this.stream.BeginRead(this.buffer, 0, this.buffer.Length, new AsyncCallback(OnRead), null);
-        }
 
 
     }
